+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 0.49.0
 
--   **Deprecated** `immutable` option in favor of `autoOptimizeInserts` option that works well in an insert-only
+-   **Deprecated** `immutable` option for `processChangeStream` in favor of `autoOptimizeInserts` option that works well in an insert-only
     or insert-heavy scenario. If you only want `insert` events, you must pass `{operationTypes: ['insert']}`.
 -   Only return `_id`s for `failedDocs` and `failedDoc` in emitted events.
 -   Bumped dependencies.